backend/requirements.txt
fastapi
uvicorn
pillow
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
requests